+++ src/components/registerSection/CardIssuerInputSection.tsx
@@ -1,7 +1,7 @@
 import * as S from '../../app.style';
 import * as Styled from './registerCardIssuer.style';
 import Label from '../composables/Label';
-import InputSection from '../composables/InputSection';
+import InputSection from './InputSection';
 import { CARD_ISSUER } from '@/constants/cardSection';
 import { forwardRef, SelectHTMLAttributes } from 'react';
 import INITIAL_CARD_ISSUER_INFO from '@/constants/allCardIssuerInfo';
